requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
asttokens==2.4.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
crashtest==0.4.1
cryptography==39.0.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
dulwich==0.21.3
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.9.0
frozenlist==1.4.1
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
ipython==8.12.3
jaraco.classes==3.2.3
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.17.3
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
p2pnetwork==1.2
packaging==23.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pickleshare==0.7.5
pigar==2.1.4
pipenv==2023.2.18
pipreqs==0.5.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pystun3==2.0.0
python-dateutil==2.9.0.post0
pyzmq==26.0.3
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tomlkit==0.11.6
tornado==6.4.1
traitlets==5.14.3
trove-classifiers==2023.3.9
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.9.4
zipp==3.15.0
